refactor(queue): merge MailModule into QueueModule, delete mail directory
- Delete MailModule and mail/ directory entirely
- Merge EnqueueEmailUseCase into queue/use-cases with direct SQSClient injection
- Delete QueueService — use case handles SQS send directly
- Remove MailModule imports from AuthModule, UsersModule, SurveysModule
- Update all imports to @/app/queue/use-cases/enqueue-email.use-case
- Update docs to reflect merged QueueModule structure
