+++ src/app/http/users/users.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { CryptographyModule } from '@/app/cryptography/cryptography.module';
-import { MailModule } from '@/app/mail/mail.module';
 import { StorageModule } from '@/app/storage/storage.module';
 import { Token } from '@/domain/entities/token';
 import { User } from '@/domain/entities/user';
@@ -25,7 +24,6 @@ import { UsersController } from './users.controller';
     TypeOrmModule.forFeature([User, Token]),
     CryptographyModule,
     EnvModule,
-    MailModule,
     StorageModule,
   ],
   providers: [
