--- src/app/mail/mail.module.ts
@@ -1,6 +1,0 @@
-import { Module } from '@nestjs/common';
-
-import { EnqueueEmailUseCase } from './use-cases/enqueue-email.use-case';
-
-@Module({ providers: [EnqueueEmailUseCase], exports: [EnqueueEmailUseCase] })
-export class MailModule {}
