+++ src/app/queue/use-cases/enqueue-email.use-case.ts
@@ -1,24 +1,26 @@
 import { SendMessageCommand, SQSClient } from '@aws-sdk/client-sqs';
 import { Injectable } from '@nestjs/common';
 
+import { Log } from '@/common/log/log.decorator';
+import { LogService } from '@/common/log/log.service';
 import { EnvService } from '@/env/env.service';
-import { SendEmailJob } from '@/shared/queue/email.dto';
-import { MessageEnvelope } from '@/shared/queue/envelope';
-
-export const SQS_CLIENT = 'SQS_CLIENT';
+import type { SendEmailJob } from '@/shared/queue/email.dto';
+import type { MessageEnvelope } from '@/shared/queue/envelope';
 
 @Injectable()
-export class QueueService {
+@Log()
+export class EnqueueEmailUseCase {
   private readonly queueUrl: string;
 
   constructor(
+    private readonly logger: LogService,
     private readonly sqs: SQSClient,
     private readonly envService: EnvService,
   ) {
     this.queueUrl = this.envService.get('EMAIL_QUEUE_URL');
   }
 
-  async enqueueEmail(job: SendEmailJob): Promise<void> {
+  async execute(job: SendEmailJob): Promise<void> {
     const message: MessageEnvelope<SendEmailJob> = {
       version: 1,
       type: 'email',
@@ -31,5 +33,10 @@ export class QueueService {
         MessageBody: JSON.stringify(message),
       }),
     );
+
+    this.logger.log('E-mail job enqueued', {
+      template: job.template,
+      to: job.to,
+    });
   }
 }
